requirements.txt
numpy>=1.19.5
pandas>=1.3.0
scikit-learn>=0.24.2
tensorflow==2.19.0
xgboost>=1.5.0
requests>=2.26.0
plotly>=5.3.1
streamlit>=1.32.0
textblob>=0.15.3
python-dotenv>=0.19.0
ta>=0.7.0
yfinance>=0.1.63
beautifulsoup4>=4.9.3
scipy>=1.7.0
optuna==3.5.0
prophet==1.1.5
statsmodels==0.14.1 